prom-file-count: include symlinks & other special files
This is needed to use the script to track open FDs for a process
as the files under /proc/12345/fd are symlinks.

Verified by manual testing that it does not change the value of any
metrics currently exported on fermium (mailman list master), the other
spot in production where this script is used.

Bug: T243634
Change-Id: I103c4067b3c6e0d631564fb37d5c52c47fc4b290