doc: stop using wikidev group, use doc-uploader group
A follow-up to I9d63e8400c9 we stopped using the wikidev
group and instead expect users to sudo to doc-uploader.

doc-uploader is both a user and a group so we can use it
instead and avoid wikidev.

Then we also don't need to include the umask change for wikidev anymore.

Bug: T237707
Change-Id: Iab82faa92b586e4aadd1f78f3c0fb779d1c0bd3a