Revert "cloud: hiera: puppetmaster: refactor hiera (for VM instances)"
This reverts commit d270782d166dac90ab35e866f8238334cee2ca0e.

Reason for revert:

The wmcs_deployment variable is set on a manifest that depends on hiera (realm).
In the original patch, the $::realm variable can't be queried if you don't know
the wmcs_deployment beforehand. There is an chicken-egg problem here that will
likely need another approach before we can rearrange all the hieradata.

Change-Id: Ib808e7b4ad35b3de56474aa02619d03e4594d3fd