Keystone: set max_active_keys for fernet tokens
We're not actually using fernet tokens yet, but setting
this in the config will let me work on staging the keys
properly before we use them -- this setting is used by
keystone-manage when managing tokens.

For the first pass we're going to have each keystone host
rotate keys once per day, and then rsync to the other host.
That means two rotations per day; the max lifespan of a token
is 7.1 days.  18 keys allows for two per day plus two for staging
plus two more in case my math is off by one.

Bug: T243418
Change-Id: Ie937a401502e100929d44b149a10f6e1f283c365