presto: add kerberos and tls support
What:
Add Kerberos and TLS support for Presto and enable it for
Hadoop test.

Why:
Presto offers HTTPS as only way to use Kerberos authentication.
By default the HTTP port is unauthenticated, so it needs to be
disabled in case authentication is required. This change adds
the support for all the config settings.

Change-Id: Icaadea46c752cae3dd16c7b930c24855c0a861f2