realm global: make the realm variable a global in labs
Currently the `$realm` variable is hiera value in labs but a global
variable, via realm.pp, in production.  This CR unifies the two
environments so that the realm variable is configured in realm.pp.
This also allows global variables that rely on the realm variable to be
available to hiera when it computes the hierarchy

Bug: T244222
Change-Id: I30f31895ccc064fad975fc50f9955fba6f0aaffd