profile::memcached::instance: add the theads parameter
What:
Add the possibility to set the number of processing threads for
memcached.

Why:
By default memcached uses only 4 threads to process requests,
and historically (due to heavy use of locks) the community and
upstream suggested to avoid values higher than 8 to avoid performance
penalties or bottlenecks. With memcached 1.5.x the scalability has
improved a lot, so this value can be raised more.

Change-Id: If9f149130088c08734ff6403ee2d3b29ec9fb283