gerrit: make db_pass configurable from (private) Hiera
Because we want to be able to use a different (readonly)
user for the temp. Gerrit test server and currently the
user is configurable but the password is hardcoded in the
passwords module in the private repo.

Before merging the existing password needs to copied from the
passwords module to private hieradata.

Bug: T243800
Change-Id: Ie01c269e836487757e8178f72abaeddd33b56192