gerrit: for db_user, remove default and replace hiera() with lookup()
follow-up to Ie01c269e83648

In general we want to move away from hiera() and use lookup()
and not have any defaults.

And in the private repo the db_pass is moved from the passwords
module to private/hieradata.  So this is to match that and make
things more consistent, so that db_user and db_pass can both be
found in the role/common Hiera, just public vs. private.

Bug: T243800
Change-Id: I2012c55e7a3575809d6c423a713146d8afab8a45