vcl: merge wikimedia-common into wikimedia-frontend
There is no reason to have "common" VCL now that we use Varnish at the
frontend layer only. Merge common and frontend VCL. Where it seemed
appropriate, complex and/or long subroutines have been left defined as
subs. In other cases we have just included the code and avoided
subroutine calls.

Move alternate-domains code to a separate file so that we can include it
before the role-specific VCL.

Bug: T241239
Change-Id: I7d04ce64cb59595d51578467d05e8c885c2b4f21