backups: Migrate db1095 backup source instances (s2, s3) to db1140
db1095 is the backup source for s2 and s3 on eqiad. It crashed
and may be unstable or, more important, have a performance impact
while its hw issues are solved.

However, backup taking cannot stop. While we have redundancy on
codfw, having a fast backup recovery path on a local dc should
always be available.

For this reason, setup db1140 (which used to only have x1 backup
source database, and was considered for the most part a "spare"
backup source for growth and incidents) with also s2 and s3.

The host had 5TB available, so disk space won't be an immediate
concern. However, while db1095 is offline, the total available
memory will be reduced, which could result in slower backup
taking.

Bug: T244958
Change-Id: Ib15c89ffac28fd33345b5b689f652a782482ae85