Add script to track OS migrations status
This script detects the role for every server running in production and
generates stats for all systems not running the latest OS(es) (it can temporarily
be two ones if we e.g. test-drive a new Debian release before it's released as
stable) as to when they are planned to be migrated, how many systems are affected
and if any systems are behind plan. A report text file is generated.

Change-Id: I7a71df642e18874e2eb50b481c0bbb217fd5c62f