otrs: Add wmfjson logs to OTRS
Why:
Use OTRS apache as a testbed to verify the validity of the approach

What:
Use the new wmfjson LogFormat in a new CustomLog directive. This
duplicates all logs in a json only file as well

Note that this is expected to double the disk usage for logs on
mendelevium. That's fine, the amount of apache logs on mendelevium is on
the order of MB.

Change-Id: I796b04f6b227a9667da5ac4e2f2bd8ca87a038b0