tlsproxy::localssl: change duplicate definition detection
tlsproxy::localssl currently has a bit of a hack to avoid multiple
defines claiming to be the default_server.  the hack defines a notify
(noop) resource which if defined twice would cause a catalogue
compilation error.  unfortunately even though notify is a noop operation
it still shows as a change in puppet reports.

This CR updates the notify resource to an exec resource setting `onlyif
=> false` which should ensure it never runs and therefore wont cause a
change however the resource is still defined so it should catch multiple
definitions

PCC: https://integration.wikimedia.org/ci/job/operations-puppet-catalog-compiler/20810

Bug: T242910
Change-Id: Iee8c8f0929efeb0b9c7c47065cd3d0a18d2e8cad