openstack::clientpackages::mitaka::buster: change notice to warning
We are attempting to report on systems that preforme a puppet change on
every run.  This class sends a notice stating that the class is not
implemented.  It seems like a warning makes more senses in this instance
further a warning does not cause a change.

i do wonder if we could remove this all together?

Bug: T242910
Change-Id: Ia1b288d226dcd442949cc7293157a07d820cc077