Include analytics-search-users in airflow::search role
Various debugging operations, such as checking yarn logs
or manually re-issuing a failing command, require being
able to run relatively arbitrary commands as the analytics-search
user. The airflow specific rights for the analytics-search user
are being dropped, and the analytics-search-users admin group, which
already provides full sudo rights to the analytics-search user on
various analytics instances, is now applied to the airflow instance.

Change-Id: I7a2ffac2a7f27be4b1c2be88470e47483b4c84ca