hiera: cloud: eqiad1: use openstack service names instead of server names
Use openstack API service names instead of the harcoded server name.
As far as I know, this is used by VMs mostly to contact the API, so there is no
special couple to specific servers anyway. For VMs in the codfw1dev deployment
this is also set to the service name.

Change-Id: I68d443137faa5098ec1e97e1449b2b541448328a
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>