cache: revert Connection:KA probe experiment on cp4026
Using 'Connection: keep-alive' for probes did not work. All backends
were marked as 'sick', and asking varnishlog to print backend_health
messages showed a 200 coming from the ats-be side, plus part of the
headers. See https://phabricator.wikimedia.org/P10459. Headers are not
printed at all in case of successful healtchecks.

Revert the experiment on cp4026 for now, but keep the code allowing to
switch between 'close' and 'keep-alive introduced in 9726e64546.

Bug: T244464
Change-Id: I3422c8c480c8d4b0b6780b745d616ade8b9c7079