cloudstore: Update the nfs_hostlist script
Wrote a script to do effectively what this is for. Then I took my features
and reasoning and put it into this script when I found it.

This optimizes much of the looping into set arithmetic and list comprehensions.
It also allows you to select hosts based on either project or mount specification.

Also, updated the URL of openstack-manager to be the newest modern one.

Bug: T224582
Change-Id: I5e595e73324694aafd44cb7f5f5bac2ddd93710f