Add CAS authentication to tendril
This enables authentication via the mod_cas Apache module. If a user accessing
tendril has an active session with the CAS SSO site, access is granted depending
on the LDAP groups membership of that user. If no active session is found for
the user, they are redirected to https://idp.wikimedia.org/login for authentication
and then Tendril can be accessed.

Enable this via a class option to be able to easily revert in case of issues;
if all works fine, the LDAP authencation Puppet code can be dropped in a future
step.

Tendril will see some changes in the future in terms of a rewrite of the
codebase, but that the authentication is agnostic to what's running behind
it, so this can simply be reused for tendril-ng. That's also why this commit
limits changes to the tendril to the minimum required (and e.g. doesn't
add type hints to the parameters etc).

Change-Id: I5c8ad4270e61738e72d2bccda70fb13b808ad1cd