httpd: Log X-Client-IP header if it exists
Why:
Tracking down external requests in our mediawiki servers without this
information can be painful as there are a large numbers of intermediate
reverse proxies, last of which is the host itself these days (because of
TLS termination

What:
Add X-Client-IP header logging in both:
* wmf LogFormat
* wmfjson LogFormat

Do so for both regardless of whether removeip is used or not. In the
latter case however, rename the old X-Client-IP wmfjson key to Client-IP
so that it doesn't conflict.

Change-Id: I994caca429990db88232082097906cd8db06eccf