toolschecker: do not verify TLS cert for k8s API checks
The 2020 Kubernetes cluster uses a TLS cert that is signed by an
internal CA. Rather that faff about to get the signing cert to verify
the TLS cert, we will just tell the check to ignore validation of the
server cert. This mirrors what has been done in the past for our etcd
checks.

Change-Id: Ifb26599d394a591c2f5853bae9912923f87a5740