configmaster: Add DNS Discovery discrepancy check
Why:
Operational changes to DNS Discovery records are not currently tracked
in puppet or any kind of git repo. They are logged into SAL, and while
SAL is good for the why and the when of a discovery record being
pooled/depooled, it just isn't helpful as a notification. Furthermore it
might not be clear to everyone what the desired state is and whether
something should be diverging from it or not

What:
Add a simple python script wrapping confctl that checks the current DNS
discovery state against a, currently hardcoded internal one. Include
some instructions on how to update that structure.

This is a first step in this direction, we might want to augment this
into something more generic that allows to diff arbitrary entries in our
discovery system

Change-Id: I9067357ca2dac2262bb4ebf5a6ed96ddb9c96a14