aptrepo/install: move https monitoring to aptrepo profile
The HTTP check does not have a URL parameter and can (should)
run on each server using these profiles. Therefore it matches
the profile for the HTTP server.

The HTTPS check though is specific to https://apt.wikimedia.org
and therefore it fits better into the APT specific profile.

Here we have the parameter to check whether we are on the primary
server and only do monitoring once.

Avoid adding the same check multiple times to Icinga now that we
will temp. have 4 install servers.

Bug: T224576
Bug: T242602
Change-Id: Ie5253814857096fccd946a20fdba64c0bf752d5a