Add profile::base::no_firewall
Now that we have almost complete coverage of profile::base::firewall I
want to add a consistency check which validates that every role has
base::firewall applied unless it's explicitly opt-out (so that we're
able to spot regressions if new roles don't use Ferm).

Add a stub profile which can be used to designate that a given role
intentionally does not use base::firewall.

Bug: T104939
Change-Id: Id8cd12210098707e9d9e3cf98af5da7b031c3388