httpbb: Tiny script to work around import path issues.
Now that httpbb is more than one file, symlinking to main.py no longer
works, because it can't import other httpbb modules without fiddling
with $PYTHONPATH or similar (which we'd rather not do).

This is only necessary because we install via git clone -- when we have
a proper debian package instead, we'll remove this.

Bug: T236699
Change-Id: I0dd06cfa773bcf9d4496f68cd8e68d347b606f17