Move all Report Updater Jobs to an-launcher1001
What:
Move the remaining Report Updater jobs from stat1006 to an-launcher1001

Why:
This is part of an effort of the Analytics team to move periodic team jobs
to a dedicated host rather than using client nodes.

This change introduces a new profile, since the mysql uses the 'stat' user,
meanwhile hadoop uses 'analytics'. In this way on an-launcher1001 all the
RU jobs will be run as user 'analytics'.
It also refactors statistics::compute to avoid stat100x specific bits that
should not be deployed on an-launcher1001 (like the /srv and /home rsyncs).

Bug: T243934
Change-Id: I5e176061b83c8f0bbc8e9a48dedfc0b4e4bfcc58