Enable CASValidateSAML for tendril
CAS was throwing 401s for all requests which were passed through the
rewrite statements for index.php. Logs were showing apparent failures
to properly detect the user who's authenticating. Luca found a fix,
though: If CASValidateSAML isn't enabled (which is the default), the
remoteUser is read from the header (and that fails due to mod_rewrite
intervening), but if CASValidateSAML is enabled, the user is previously
parsed and then authentication succeeds.

This also involves changing to SAML 1.1 (and thus changing the validation
URL)

Add a new option to profile::idp::client::httpd and enable it for
tendril.

Change-Id: I985f05c6f60a4a68c8e96f047e0adab287b98d95