Add an-launcher1001 to the list of statistics servers
What:
Add an-launcher1001 to the statistics_servers hiera list.

Why:
The host is not really a stat host but it is taking over jobs
from stat1006 and stat1007 (Report Updater timers) and it needs to
be able to publish the results to thorium via rsync. There will be
a follow up refactoring to separate concerns between an-launcher and
stat boxes.

Bug: T243934
Change-Id: I42538746162fa53644973a1b7d75700188ac9f44