Deployment group audit
This patch attempts to clean up the Wikimedia deployment group
per the audit steps within the associated Phab task.  Basically:
remove the deployment right for individuals who have not deployed
anything within two years of 2020-02-25.  I've tried to avoid
removing new WMF employees and folks who were recently added,
namely: mstyles, zpapierski, clarakosi, dpifke, niedzielski.
Also, the audit script may not be _perfect_ as it is attempting
to process the on-wiki SAL as best as it can.

I'll plan to add removed folks to this patch.  If you still
require deployment rights, please comment on the patch or Phab task
with a brief explanation as to why, and it most likely shouldn't
be a problem to add you back.  But please be honest about
your actual _need_ for the deployment right.

Notes:
* kaldari, awjrichards, dr0ptp4kt confirmed removal
* moving musikanimal from deployment to restricted per T237696#5923923
* restoring phuedx per T237696#5922925
* restoring samwilson per T237696#5925494

Bug: T237696
Change-Id: Idbf37f207e788176cdd25415ec6554bd865038a5