Unroll analytics Partman configs
There's no real reason all Ganeti Partman configs are assigned in a single
line; it's hard to read and error-prone (installs were recently broken
as a superfluous | was added in the long host globbing). Also switch
to starred notation, which avoids the need to change netboot.cfg the
next time a server is added or removed.

Bug: T156955
Change-Id: I862c22a2968020c8705ecb131ab57cd1cb4d4ed1