Unroll Partman configs for Ganeti-based clusters
There's no real reason all Ganeti Partman configs are assigned in a single
line; it's harder to read and error-prone (installs were recently broken
as a superfluous "|" was added in the long host globbing). Also switch
to starred notation, which avoids the need to change netboot.cfg the
next time a server is added or removed.

(etcd1001-1003 are being removed in T239835, dropped them already)

Bug: T156955
Change-Id: Ia7a7b3be4e67533680fd4a48078a005217fbe9e8