parsoid: change cluster of parsoid-test machines from misc to parsoid
The cluster is currently the default "misc" on scandium.

But $cluster is used to set $_SERVER['SERVERGROUP'] in
profile::mediawiki::httpd which in turn is then used
in MediaWiki's CommonSettings.php to check whether on a parsoid
server or not.

This is to make scandium a member of the 'parsoid' cluster as well
so we can avoid hacks involving a hard-coded hostname.

per IRC and:

https://gerrit.wikimedia.org/r/c/operations/mediawiki-config/+/575336/2/wmf-config/CommonSettings.php#4000

Change-Id: Ie43e611d10a937291e3b3a2c63f296d1831f54e8