prometheus-mysqld-exporter: Fix port of multisource host
This was a weird bug that got unnoticed for a long time:

For multisource hosts, get_socket() was called twice, which ended
up causing to consider multisource hosts as multi-instance, and
configuring them to use 19104 port instead of 9104.

This went unnoticed because there was a manually-corrected definition
of multisource hosts (labsdb1009-12) with a different name, labs
instead of labsdb), unpuppetized. This was only noticed when
some graphs showed multiple versions of these hosts.

The duplicate files have now been deleted and this patch will
restore the port to the right value.

Change-Id: I085a4ced3a7c7d8a50eedf523187a1f7a05801e4