prometheus::ops: collect envoy stats from all servers
Up to now, we were only collecting envoy stats from k8s-enabled
services, but now we want to collect the envoy stats from all servers.

We will limit the amount of metrics we collect in order not to
overflow prometheus.

Also fix the class name we're collecting installations with, and the
job name.

Change-Id: I6afe8bb5b50224421cea3d926b92f17360a06c26