grafana: remove the distinction between readonly and admin domains
The grafana-labs server had two endpoints, one that automatically
logged in with a dummy user (the 'readonly_domain') and one that
required a login with a valid account (the 'admin_domain')

In prod we've only had one domain for quite a while, this domain
was inaccurately labeled 'readonly_domain' which allows
reading of metrics by logged-out users and editing by logged-in
users.

This patch makes the grafana-labs endpoint resemble prod.  The hiera
params are collapsed down to just one domain called 'domain' and
grafana-labs is now backed via an ldap login, just like prod.

This probably breaks grafana editing for some edge-case users
who aren't in the 'nda' group in ldap -- it's probably worth it
to prevent every single user with a wmcs login from editing
grafana dashboards.  If need be we can add a special-purpose
ldap group named wmcs_ldap_admins.

Bug: T246508
Change-Id: Iaa6be2cdf100b0ccd0b80d2fb3a5d2cb4eea7aee