Allow analytics client nodes to set user.slice global limits
What:
Refactor systemd::slice::all_users to allow global limits
via override of user.slice.

Why:
The Analytics client hosts (like stat100X) are usually very crowded and
nobody really cares about resource consumption of processes/scripts/etc..
We tried with per-user limits, but it doesn't work since when 2/3 users
start to crunch data with heavy processes the host goes always in trashing
mode. This change is meant to add a global limit for all the users, to
see if any difference can be made. The downside is that user-0 will also
be affected, but I'd be inclined to try anyway since the current status quo
is not sustainable (too many alarms, user complaining about user limits
breached, etc..).
It should be a no-op for all the WMCS use cases, and only a change for
Analytics hosts.

Change-Id: If31bfce7641925cd09395b5cca7f026b3668d278