cdh::hive: remove DBTokenStore from hive-site.xml config
What:
Remove DBTokenStore references from hive-site.xml

Why:
The DBTokenStore setting was helpful to address the issues
described in https://gerrit.wikimedia.org/r/#/c/operations/puppet/+/562474,
but it turned out to be a problem with Hive 2.2.3 and Oozie 4.3.
The setting seems to be related to a High Availability Metastore,
something that we are not doing. Until recently we had used Zookeeper
settings in the hive-site.xml, that I suspect were also enabling
some HA code parts of Hive. I'd like to try the previous settings
that we have and see if the Kerberos sporadic errors come back.

Bug: T244499
Change-Id: Ia89411dc0287b7fbbe51f8a68c9ae6c56d55dfef