ferm: Add status check
currently there is no way to check if the correct iptables rules have
been loaded into the kernel.  This CR adds a script which compares the
output of `iptables-save` with the output of `ferm -nl
/etc/ferm/ferm.conf` to ensure the iptables rules configured in ferm
are present in the iptables.

conflicts with https://gerrit.wikimedia.org/r/c/operations/puppet/+/573335

Bug: T206951
Change-Id: I6447500096eb55534d69bd0c7db1a5b3fa523272