installserver/apt_repo: add homedir parameter, move dirs to Hiera
follow-up to I7722abbb5747a9b8

We have to add the "homedir" parameter to the current and new APT
repo roles because it defaults to /var/lib/reprepro as used on
releases servers and we need the GPG keys in /root.

Also move both basedir and homedir parameters to Hiera and use lookup.

Bug: T224576
Change-Id: I1b032afc6ebcdb0d43aadf6ee3917bc902dee021