openstack: serverpackages: queens: stretch: introduce some apt pinnings
In order to install certain packages for openstack queens, we need to pull
packages from the standard stretch-backports repository (they are not included
in our special mirror, but in the official debian archive for stretch-bpo).

With this pinning, apt-get will be smart enough to install the packages from
proper repositories.

Actually, this might be only needed when installing nova, so it would make no
sense to have this pinning in other servers where nova is not installed, but
we might soon discover that we need pinning for other openstack services too,
so I decided to introduce this pinning facility here for simplicity (instead of
in openstack::nova::common::queens::stretch).

Bug: T246671
Change-Id: I954c0f35f4104c6fd43e8112c6fc8f4d1d2351ee
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>