mediawiki::maintenance: force removal of directories
Currently when the mediawiki::maintenance profile runs on the failover
site it has ensure set to absent.  this causes puppet to attempt to
remove some directories which fails with the following message.

  Notice: /Stage[main]/Mediawiki::Maintenance::Cirrussearch/File[
  /var/log/mediawiki/cirrus-suggest]: Not removing directory;
  use 'force' to override

This CR sets the file resource to use force and ensure the directory
gets removed.  This will also mean anything within that directory e.g.
old logs etc will also be deleted.  it may be better to just always
`ensure => directory` for the directories regardless of which site is
live

I also change the resource defaults to use use per-resource default attributes

Bug: T242910
Change-Id: I64bfdec3eabf9cd0a50611b83a57f5334d606ea9