Ensure readability settings for home dirs of Analytics clients
What:
Ensure readability settings (chown/chmod) for home directories
of users on Analytics clients belonging to groups with access to
PII data.

Why:
On the Analytics client nodes multiple user groups can coexist.
The analytics-privatedata-users one is particularly sensitive since
it allows users that belongs to it to pull PII data from Hadoop.
If users are not careful enough to set good permissions for those files
(and experience tells us that this is the case sadly), then it means
exposing sensitive data to potentially non-autorized users.
This change adds a script that simply chmods/chown the home directories
of users in analytics-privatedata-users to something reasonable.
The script can be extended in the future to more groups if necessary.

Bug: T243934
Change-Id: Iaf9b465c9530eca3de35be1867c2c3b24e29f3a1