keepalived: add initial module and toolforge profile
This initial keepalived module provides support for managing a virtual ip address
(VIP) on a cluster of nodes. The virtual router redundancy protocol (VRRP) is
being used to automatically configure and discover the available MASTER and
BACKUP nodes in the cluster.

Keepalived is being used by Toolforge to manage a high availability service
address across the new Elasticsearch cluster members.

Bug: T236606
Change-Id: I7cdbdbd09df8378db7db0e6c2e02b1c3299cb6af