installserver: ensure interface for DHCP server is configured
On jessie the DHCP server simply runs with the default
"INTERFACES=" but on stretch and buster it fails if no
interfaces are set explicitly.

Also it's now INTERFACESv4= and INTERFACESv6=.

None of our subnet definitions have v6 so it's just v4 for now.

Use file_line to ensure the INTERFACESv4 line has an interface
in the defaults file installed by the package.

Put it into a temp. "if buster" to avoid any changes on the
existing servers until they are replaced.

Also add missing lookup() and Hiera key for the
$ensure_service variable, follow-up to I22b8bb22d9c6d74.

Bug: T224576
Change-Id: I76471a0831bedf86cf924ffe916e992f16e4ffd4