httpbb: Replace apache-fast-test with httpbb in deploy_apache_change.
This replaces the testing tool (and removes the $control variable, since
httpbb doesn't need one) and also moves the script from
profile::mediawiki::web_testing (where apache-fast-test lives) to
profile::httpbb. Presently those profiles are on the same hosts, so that
move is a no-op.

In a followup, we may want to change the usage of deploy_apache_change,
to allow for more than one test file to be passed as an arg. For now,
though, the only change was to make that arg optional, which is
backwards-compatible.

Change-Id: Ife9b8b74d85658c703430a11e6719ab142f15ff5