profile::superset: allow to deploy Presto TLS CA
What:
Add the possibility to deploy the TLS root CA of
a Presto cluster in Superset.

Why:
Presto uses self-signed TLS certificates, and Presto
works only with https with Kerberos. Superset needs to
have the root CA certificated deployed on the host.
The presto client profile was another option to use but
it requires packages like the opendjk, not really needed
for this use case.

Bug: T239903
Change-Id: Iaa50dcc6613476f89d051036dbee1bed7db58231