LVS: add alert on CPU saturation, which causes pkt drops
Looking at historical data this is a very good signal.
Make sure you're logged into Grafana and then PTAL:
https://w.wiki/HQo and https://w.wiki/HQp

Also clean up an already-resolved TODO comment.

This is a re-creation (with followup fixes) of I4e9f93c1c4.
It includes the fix from I065d15d0, and also corrects a {
that was erroneously written as a (, causing the errors cited
in Iebdd909c5.

This reverts commit c0fec4239f6772383db6d3c8510dcca56e382dfe.

Change-Id: I60467b10ef7e8ced81f0198321a20df73add52df