envoyproxy: run validation as user envoy
Quite surprisingly, when running in mode "validate" envoy will still
create its logfiles. Given it's running as root, that means the files
will be created as root:root and thus won't be usable when we run the
service as user envoy.

Resolve this bootstrapping issue by running the verification step as
the same user.

Also fix some black-conformity in formatting.

Change-Id: I6faeca091553a6df3a780f7fd13c5eb500e2db48