ATS: Disable parent proxies on ulsfo
Now that KA is enabled between ats-tls and varnish-frontend it should
be safe to disable parent proxies and just hit one varnish-frontend
port. This obviously will benefit the connection reuse between the
two layers and we should be safe as long as the connection reuse
stays over 8 requests per connection.

Bug: T244464
Change-Id: Ie3f389dfd4cecf12957e804f889e2997b687e352