Add xmldumps to stat100[4,5]
What:
Add the configuration on labstore nodes and on stat100[4,5] hosts
to mount xmldumps mountpoints.

Why:
We are trying to spread users on multiple Analytics clients, since most
of them are clustered on stat1007 (since stat1006 is not a hadoop client
for the moment). Some of these users would be happy to migrate if the xmldumps
mountpoints would be available on stat100[4,5] too. We are not foreseeing
more traffic to labstore nodes, just the load spread on multiple clients.

Bug: T243934
Change-Id: I74588c2d12243a9fdca5f7971e8ede5bfdf86df4