offboard-user: update script so that it can traverse subgroups
The acl*securty group has been refactored so that it is a parent
container to a collection of acl*security_$department subgroups.  The CR
updates the off-boarding script to first get a list of acl*security
subgroups and add them to the list of privileged groups we should check
for when off-boarding users

This CR also removes `acl*security` from the list of privileged groups.
the api can no longer removes members from the group with the API
producing the following error message:

  phabricator.APIError: ERR-CONDUIT-CORE: Transaction with key "0" has
  invalid type "members.remove". This type is not recognized. Valid types
  are: parent, milestone, space, name, description, icon, color, slugs, view,
  edit, join, custom., subtype, mfa.

My suspicion is that the acl*securty project is no longer a group that
can have members signed directly but this needs to be confirmed by
chasemp or possibly someone with a better understanding of phabricator
groups

Bug: T245771
Change-Id: Id321c594457f4d1e3ed7e95106fd5f91977963ae