mediawiki: Change php-wmerrors channel from "fatal" to as "exception"
MediaWiki previously used both of these depending on how it found
the fatal error. On the php-wmerrors side here we couln't distinguish
between the two so we just picked one ("fatal"), and our monitoring
queries and dashboards always look for both as such.

As part of T247113, I'm proposing we unify these two channels
and I proposed "exception". This updates the php-wmerrors handler
to follow suit.

I'm also changing the `caught_by` field to spell out the name of
"php-wmerrors" in full, and avoid parenthesis which can sometimes
make Logstash queries more difficult than need be.

Bug: T247113
Change-Id: I564e9cf15396d630ab3b969b5d66d30f494054ab