Queens keystone: add a hack for utf8 decoding in already-hacked ldap handler
The Rocky version of this file works for direct keystone auth, but it fails
for any case where an external middlewhere (e.g. in glance) tries to authenticate.
In that case we still get a utf8 user_id which python ldap chokes on.

This hack conditionally decodes the user_id (if it's not already
a unicode string).  That doesn't break the
straight-from-the-keystone-api case and fixes the glance case.

I'm not ready to submit this upstream yet; probably Rocky
contains fixes somewhere upstream in the API to decode things
properly.  This hack lets us get a bit further with Queens,
and with luck this is all taken care of in Rocky.

Bug: T247050
Change-Id: I4d08aee639177ea0eaea523348abd873d299aadb