jupyterhub: refactor user authentication for posix groups
What:
Handle the use case of a allowed posix group, configured via hieara,
that is not present on the system.

Why:
The gpu-users group is a name that we configure in puppet, but that
refers to the 'render' posix group on the hosts. This means that
jupyterhub will fail checking raising a KeyError, and finally
returning a 500. This change unfolds the list comprehension to
handle the KeyError use case.

Bug: T245179
Change-Id: I874343c9b5e9cb280855c373fe14333b2ceb20cb