jupyterhub: fix link to frozen-requirements.txt for user creation
What:
Point to a more specific frozen-requirements.txt file when creating
venvs for users.

Why:
During the migration to Buster we are going to keep/test a different
version of Jupyterhub. In order to do so, the repository ships with
more than one frozen-requirements.txt, so the related python code
needs to be aware of it.

Bug: T245179
Change-Id: I571c8d3ee0e6d551276f8fd603be126e3fc6576e