jupyterhub: use http_proxy pid file for Buster
What:
Use a new option for Debian Buster related to the http proxy's
pid file.

Why:
On Buster we deploy a more up to date version of Jupyterhub that
requires a pid file for the HTTP proxy. Since the unit is running
with PrivateTmp is should be relatively safe/good to just use it,
rather than other places like /run.

Bug: T245179
Change-Id: Ia3feb2158e80012d76ccfabd2c6e48eb0ec2e861