statistics::mysql_credentials: use require instead of defined
What:
Replace the 'if defined' guard with an explicit require.

Why:
For some reason that I can't explain, on stat1004 the analytics-priv-data
mysql credentials are deployed by puppet, but not on stat1005. They
are configured with the same role, but they have different OS
(Stretch vs Buster). I would like to try an explicit require to
see if this is a case of silent failure due to how puppet orders
resources to apply.

Bug: T243934
Change-Id: I66c69c6ca9254a4915a192f17f4a5949dd425b15