admin: deprecate statistics-users
What:
Deprecate the statistics-users posix group.

Why:
The Analytics team is trying to simplify the number of posix groups
to get access to client nodes. All the users in statistics-users
are already in other groups (analytics-privatedata-users,
statistics-privatedata-users, researchers) except brion and mhurd.
The change adds analytics-users to role::statistics::cruncher (stat1006)
and brion/mhurd to it, to keep the same level of access.

Bug: T246578
Change-Id: I8df3734602f1edd0943b8f2522bb3328a03114c8