bacula: Increase max total size of backups to 35 TB
Old number of volumes (60) combined with the per-size volume
(500GB), limited the total size of backups, which may be used up
after retention was increased from 30 to 90 days (potentially).

Both Databases and production have currently 35 TB of available
space, which could increase in the future to 50 TB with the
unallocated space.

For that, increase max number of volumes from 60 to 70.

Bug: T238048
Change-Id: Idb3ace38597b56a5c5f98847e82915d17cac8fe3