admin: deprecate statistics-privatedata-users
What:
Deprecate statistics-privatedata-users.

Why:
The Analytics team is trying to reduce the number of posix groups
available. In this case, only two users of statistics-privatedata-users
are not in analytics-privatedata-users or researchers, namely
legoktm and risler. Moving them to researchers that provides a similar
level of access (basically Mariadb wiki replica access).

Remove also a reference of stastistics-admin in offboard-user.py
that is already deprecated.

Bug: T246578
Change-Id: Id2512a1f87c05819a04ec066fa94329ec6c434bf