check_puppet_run_changes: Ensure we exit correctly with critical
Currently the script checks for warnings before critical which means
the critical logic will never trigger.  update the script to ensure we
first check for a critical state then a warning state

the warning and critical states are currently somewhat arbitrary, with
the check entering a warning state if one host matches the test criteria
and critical if 5 hosts match the test criteria.

Change-Id: I43184b3c76d905629407287cc25cf662aaa95e06