Move stat1006 to role::statistics::explorer
What:
Deprecate role::statistics::crucher and assign role::statistics::
explorer to stat1006.

Why:
Another step towards the unification of configuration for the stat100x
hosts.
Notable step is the fact that the 'researchers' group is deployed to
stat100[4,5] too, in order to allow people in there to connect to the
mariadb wiki replicas from any stat box.

Bug: T243934
Change-Id: Ic704cfff73c2c426866c21f0d409652c00a81d92