envoy: check for runtime variables set for a long time
Runtime variables are an useful way to change temporarily the
behaviour of envoy. We should though clean them up afterwards with a
daemon reload, and not rely on them to configure the fleet.

So create an nrpe check that reports back the presence of modified
runtime variables if they persist more than 40 minutes or so.

Bug: T247387
Change-Id: I3688a9837fa5024a77220f10a5ba0280e27aab66