devtools (cloud): rm beta hostnames from scap dsh groups
There won't be actual deploying from the devtools
project to the deployment-prep ('beta') project.

Therefore remove the 'beta' hostnames to avoid
confusion.

It is unrelated to the actual beta project
and values were just copy/pasted here to test the
deployment_server role which uses scap and needed
the dsh groups.

Change-Id: I76a2aae68d7eaf41e583fb801dbb7f5689a9906b