profile::kerberos: extend ticket lifetime from 10h to 48h
What:
Set max_lifetime to 48h in the Kerberos KDC's config.

Why:
Currently a principal authenticating to kerberos has to
kinit every 10h to refresh the credentials, for a maximum of
7days of renewable lifetime. This causes some troubles to
data analysts during they day job, so it was asked to extend
the time before each kinit to 48h. The downside of this change
is that in case of a principal/user compromised, the KDC will trust
a ticket for (worst case) 2d more before rejecting it. This
shouldn't be a real concern in our setting since 1) we'd block
ssh sessions 2) we could delete the principal as well from the KDC.

Change-Id: I4229e6526e097baf1ed8ff12603888c9d6912c38