puppetmaster: add parameter to control ssl_verify_depth
I'm currently exploring using an external root CA to issue the puppet CA
certificate as an intermediate.  In this configuration apache needs to
be configured to verify additional leafs.

This CR allows one to configure a puppet master with SSLVerifyDepth of 1
or 2.  We currently don't have plans for more levels of intermediated
certificates so this should give us the flexibility required.

Hosts: re:puppetmaster[12].*wmnet

Bug: T247509
Change-Id: I74ad4fcc3c88b9916893af98819e459ae007f3a3