ELK7: require disktype "hdd" for new indices
Initially let's create new indices on disktype "hdd" only. This is to avoid
unwanted shard relocation as SSD hosts are added to the cluster.

Once the SSD hosts have joined the cluster, follow-up with a change to
require disktype "ssd" for new indices.

Along with this will manually need to require disktype "hdd" on
the existing ELK7 indices.

Bug: T247376
Change-Id: I7e76c64eb4eb520f6fc2954d4ce954cfb7ec2b3c